2024-04-02 review 기능 제외 모두 merge완료.
goods/get.jsp -> 쇼핑카트 , 구매로직 구현 완료
->쇼핑카트는 로그인한 사용자만 사용할 수 있도록함

----------------------
해야할 일

goods/get.jsp  프론트 수정 필요
review 기능 merge필요

+++ src/main/java/org/zerock/service/CustomOrderImpl.java
@@ -124,9 +124,15 @@ public class CustomOrderImpl implements CustomOrderService {
 				
 				for(ShoppingCartVO element : cartList) {
 
-					int result = mapper.insertOrderd_goods(""+orderno, element);
-					mapper.deleteCartElement(element.getMember_seq(), element.getCart_no()); //장바구니 항목없애기
 					
+					int result = mapper.insertOrderd_goods(""+orderno, element);
+					if(element.getCart_no().equals("no_cart")) {
+						
+						
+					}else {
+						
+						mapper.deleteCartElement(element.getMember_seq(), element.getCart_no()); //장바구니 항목없애기
+					}
 						
 				}	
 			}	
@@ -144,13 +150,14 @@ public class CustomOrderImpl implements CustomOrderService {
 		if (orderno>0) {
 			
 			dto.setOrderno(""+orderno);
-			int insertResult = mapper.insertOrderNomember(null);
+			int insertResult = mapper.insertOrderNomember(dto);
 			
 			if(insertResult>0) {
 				
 				for(ShoppingCartVO element : cartList) {
 
 					int result = mapper.insertOrderd_goods(""+orderno, element);
+					
 									
 				}	
 			}
